refactor: add wrapper for API changes, adjust some height items
